# gestion-soutenance-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Soutenances",
        "description": "Documentation OpenAPI pour le backend Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/etudiants/{etudiant}/rapports": {
            "get": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Lister les rapports d'un étudiant",
                "operationId": "2d4c92b838a7c9aac0b5fab03389d10a",
                "parameters": [
                    {
                        "name": "etudiant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des rapports"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Dépôt de rapport par étudiant",
                "operationId": "cb9062e939a8e3cb86bdb1a2f7afa71f",
                "parameters": [
                    {
                        "name": "etudiant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "date_depot"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "date_depot": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "commentaire": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rapport créé"
                    },
                    "400": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/juries": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister les juries",
                "operationId": "742f25f4fe4d930e29cc345de477347e",
                "responses": {
                    "200": {
                        "description": "Liste des juries"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer un jury",
                "operationId": "a82970fa5418f991fc2b0e697cf4cdf6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "president_id",
                                    "rapporteur_id",
                                    "encadrant_id",
                                    "examinateur_id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jury créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/juries/{jury}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Mettre à jour un jury",
                "operationId": "6eade27ab0d6a937d07b83fb5d86ef63",
                "parameters": [
                    {
                        "name": "jury",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un jury",
                "operationId": "ba624e52e82d529369f939cb85fce50e",
                "parameters": [
                    {
                        "name": "jury",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/professeurs/{professeur}/rapports": {
            "get": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Lister les rapports des étudiants assignés (encadrant ou rapporteur)",
                "operationId": "4582b2a5de7d8fce29925a058ca522e0",
                "parameters": [
                    {
                        "name": "professeur",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des rapports"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/rapports/{rapport}/remarques": {
            "get": {
                "tags": [
                    "Remarques"
                ],
                "summary": "Lister les remarques d'un rapport",
                "operationId": "c01bb2d432ca27b16b1b280f67ba3c03",
                "parameters": [
                    {
                        "name": "rapport",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des remarques"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Remarques"
                ],
                "summary": "Ajouter une remarque sur un rapport",
                "operationId": "0e71d10818c2da41f78f3259acf7b910",
                "parameters": [
                    {
                        "name": "rapport",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contenu"
                                ],
                                "properties": {
                                    "contenu": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Remarque créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Saisir uniquement le jeton d'accès",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}